[Calib] Move public headers to interface directory
