+++ CalibMuon/DTDigiSync/plugins/BuildFile.xml
@@ -1,0 +1,5 @@
+<use name="FWCore/PluginManager"/>
+<use name="CalibMuon/DTDigiSync"/>
+<library file="*.cc" name="CalibMuonDTDigiSyncPlugins">
+  <flags EDM_PLUGIN="1"/>
+</library>
